New Layout component below FilterHub
