+++ omero_parade/static/omero_parade/css/parade.css
@@ -8,12 +8,18 @@
 }
 
 /* Top-level React element */
-.plateContainer {
+.reactContainer {
   position: absolute;
   top: 0;
   bottom: 0;
   left: 0;
   right: 0;
+  display: flex;
+  flex-direction: column;
+}
+
+.plateContainer {
+  flex-grow: 1;
   display: flex;
   flex-direction: column;
 }
